terminando api

+++ lojamusical/src/main/java/com/lpweb/lojamusical/model/base/EntityBase.java
@@ -9,13 +9,13 @@ import javax.persistence.MappedSuperclass;
 public class EntityBase {
 	
 	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected Long id;
+    protected Integer id;
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
